try making links branch for history

+++ scripts/gen_md.py
@@ -63,12 +63,15 @@ def generate_markdown_from_json(morning_run, json_dir, output_dir):
 
         # Ensure the output directory exists
         os.makedirs(output_dir, exist_ok=True)
-        
+        category = category.replace(" ", "-").lower()
         # Create the output markdown file with UTF-8 encoding
         mor_prefix = "mor_" if morning_run else "eve_"
         output_file_path = os.path.join(output_dir, f"{mor_prefix}{category}_{today_date}.md")
-        with open(output_file_path, 'w', encoding='utf-8') as output_file:
-            output_file.write(markdown_content)
+        if not os.file.exists(output_file_path):
+            with open(output_file_path, 'w', encoding='utf-8') as output_file:
+                output_file.write(markdown_content)
+        else:
+            print(f"Skipping {output_file}, exists...")
 
 # Set the paths to your json directory and output markdown directory
 json_directory = 'tmp_json'
